# src/ThreatModelForge.Api/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ThreatModelForge.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/model/analyze": {
      "post": {
        "tags": [
          "Model"
        ],
        "operationId": "AnalyzeModel",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TmForgeModelDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FindingDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/model/analysis": {
      "post": {
        "tags": [
          "Model"
        ],
        "operationId": "RunAnalysis",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TmForgeModelDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/v1/model/analysis-document": {
      "post": {
        "tags": [
          "Model"
        ],
        "operationId": "DescribeAnalysis",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TmForgeModelDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisDocumentDto"
                }
              }
            }
          }
        }
      }
    },
    "/v1/model/threats": {
      "post": {
        "tags": [
          "Model"
        ],
        "operationId": "GenerateThreats",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TmForgeModelDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ThreatDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/model/threat-register": {
      "post": {
        "tags": [
          "Model"
        ],
        "operationId": "DescribeThreatRegister",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TmForgeModelDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreatRegisterDto"
                }
              }
            }
          }
        }
      }
    },
    "/v1/model/merge": {
      "post": {
        "tags": [
          "Model"
        ],
        "operationId": "MergeModels",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergeResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/v1/model/compare": {
      "post": {
        "tags": [
          "Model"
        ],
        "operationId": "CompareModels",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCompareRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelCompareResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/v1/model/layout": {
      "post": {
        "tags": [
          "Model"
        ],
        "operationId": "LayoutModel",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayoutRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LayoutResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/v1/model/export/tm7": {
      "post": {
        "tags": [
          "Model"
        ],
        "operationId": "ExportModelTm7",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TmForgeModelDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/model/convert": {
      "post": {
        "tags": [
          "Model"
        ],
        "operationId": "ConvertModel",
        "parameters": [
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TmForgeModelDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/model/read": {
      "post": {
        "tags": [
          "Model"
        ],
        "operationId": "ReadModel",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileContentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TmForgeModelDto"
                }
              }
            }
          }
        }
      }
    },
    "/v1/model/preflight": {
      "post": {
        "tags": [
          "Model"
        ],
        "operationId": "PreflightModel",
        "parameters": [
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileContentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreflightResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/v1/model/manifest": {
      "post": {
        "tags": [
          "Model"
        ],
        "operationId": "ApplyManifest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManifestRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/v1/model/report": {
      "post": {
        "tags": [
          "Report"
        ],
        "operationId": "ReportModel",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TmForgeModelDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/model/analysis-report": {
      "post": {
        "tags": [
          "Report"
        ],
        "operationId": "AnalysisReport",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TmForgeModelDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/detect": {
      "post": {
        "tags": [
          "Formats"
        ],
        "operationId": "DetectFormat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileContentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormatDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalysisDocumentDto": {
        "type": "object",
        "properties": {
          "schema": {
            "type": "string"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "model": {
            "$ref": "#/components/schemas/AnalysisIdentityDto"
          },
          "analyzer": {
            "$ref": "#/components/schemas/AnalysisIdentityDto"
          },
          "rulePacks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RulePackInfoDto"
            }
          },
          "findings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnalysisFindingDto"
            }
          },
          "diagnostics": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "AnalysisResultDto": {
        "type": "object",
        "properties": {
          "findings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FindingDto"
            }
          },
          "threats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ThreatDto"
            }
          },
          "rulePacks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RulePackInfoDto"
            }
          },
          "diagnostics": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "FileContentDto": {
        "type": "object",
        "properties": {
          "contentBase64": {
            "type": "string"
          },
          "formatId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "FindingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "severity": {
            "type": "string"
          },
          "ruleId": {
            "type": [
              "null",
              "string"
            ]
          },
          "message": {
            "type": "string"
          },
          "elementIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "LayoutRequestDto": {
        "type": "object",
        "properties": {
          "model": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TmForgeModelDto"
              }
            ]
          },
          "page": {
            "type": [
              "null",
              "string"
            ]
          },
          "options": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LayoutOptions"
              }
            ]
          },
          "positions": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/LayoutElementDto"
            }
          }
        }
      },
...
      "ManifestRequestDto": {
        "type": "object",
        "properties": {
          "manifest": {
            "type": [
              "null",
              "string"
            ]
          },
          "force": {
            "type": "boolean"
          }
        }
      },
...
      "MergeRequestDto": {
        "type": "object",
        "properties": {
          "base": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TmForgeModelDto"
              }
            ]
          },
          "ours": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TmForgeModelDto"
              }
            ]
          },
          "theirs": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TmForgeModelDto"
              }
            ]
          }
        }
      },
      "MergeResultDto": {
        "type": "object",
        "properties": {
          "merged": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TmForgeModelDto"
              }
            ]
          },
          "conflicts": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/MergeConflictDto"
            }
          }
        }
      },
...
      "ModelCompareRequestDto": {
        "type": "object",
        "properties": {
          "baseline": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TmForgeModelDto"
              }
            ]
          },
          "proposed": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TmForgeModelDto"
              }
            ]
          }
        }
      },
      "ModelCompareResultDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "changes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelReviewChangeDto"
            }
          },
          "findingsAvailable": {
            "type": "boolean"
          },
          "unchangedFindings": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "diagnostics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentDiagnostic"
            }
          }
        }
      },
...
      "ThreatDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "ruleId": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "categoryId": {
            "type": [
              "null",
              "string"
            ]
          },
          "categoryName": {
            "type": [
              "null",
              "string"
            ]
          },
          "stride": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": "string"
          },
          "mitigation": {
            "type": [
              "null",
              "string"
            ]
          },
          "severity": {
            "type": "string"
          },
          "priority": {
            "type": [
              "null",
              "string"
            ]
          },
          "references": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "elementIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "interaction": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "justification": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "source": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "string"
            }
          },
          "manual": {
            "type": "boolean"
          }
        }
      },
      "ThreatRegisterDto": {
        "type": "object",
        "properties": {
          "manual": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "persistedGenerated": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "currentGenerated": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "staleGenerated": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "indeterminateGenerated": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "unavailableRuleIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ThreatRegisterEntryDto"
            }
          },
          "diagnostics": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rulePacks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RulePackInfoDto"
            }
          }
        }
      },
...
      "TmForgeModelDto": {
        "type": "object",
        "properties": {
          "schema": {
            "type": [
              "null",
              "string"
            ]
          },
          "version": {
            "type": [
              "null",
              "string"
            ]
          },
          "metadata": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MetaInformation"
              }
            ]
          },
          "elements": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/TmForgeElementDto"
            }
          },
          "flows": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/TmForgeFlowDto"
            }
          },
          "diagrams": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/TmForgeDiagramDto"
            }
          },
          "analysis": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TmForgeAnalysisDto"
              }
            ]
          },
          "threats": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ThreatStateDto"
            }
          }
        }
      }
    }
  },
...
}